ocr-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
numpy>=1.26.0
Pillow>=10.0.0
PyMuPDF>=1.24.0
paddlepaddle==3.2.2
paddleocr
